requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
picamera2==0.3.12
RPi.GPIO==0.7.1
numpy==1.24.3 